+++ src/Console/ProgressBar/ProgressBarException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Solital\Core\Console\ProgressBar;
+
+class ProgressBarException extends \Exception
+{
+    
+}